Split preemptions unit tests.
